Rename resizeable to resizable for consistency
