+++ ui/src/pages/namespace/Gateway/Info/index.tsx
@@ -1,10 +1,10 @@
 import GatewayInfo from "./GatewayInfo";
 import OpenAPISpec from "./OpenAPISpec";
-import ResizeablePanel from "~/util/resizablePanel";
+import ResizablePanel from "./ResizablePanel";
 
 const InfoPage = () => (
   <div className="flex grow flex-col gap-y-4 p-5 w-full">
-    <ResizeablePanel leftPanel={<GatewayInfo />} rightPanel={<OpenAPISpec />} />
+    <ResizablePanel leftPanel={<GatewayInfo />} rightPanel={<OpenAPISpec />} />
   </div>
 );
