+++ ui/src/pages/namespace/Gateway/Info/useResizeDrag.tsx
